requirements.txt

pandas==2.0.3
redis==5.0.0
config==0.5.1
openai==0.27.7
pathlib==1.0.1
tiktoken==0.4.0
supabase==1.0.3
uvicorn==0.23.1
fastapi==0.100.0
pydantic==1.10.11
langchain==0.0.284 
python-dotenv==1.0.0
